+++ src/pages/todo/TodoMainPage.tsx
@@ -1,14 +1,14 @@
 import CustomCalendar from '../../components/todo/main/Calendar';
 import AddTodo from '../../components/todo/main/AddTodo';
-import TodoHeader from '../../components/todo/main/TodoHeader';
 import MainLayout from '../../layout/MainLayout';
 import * as S from '../../styles/todo/TodoMainPage.style';
+import Header from '../../components/common/Header';
 
 const TodoMainPage = () => {
 	return (
 		<MainLayout>
 			<S.Container>
-				<TodoHeader />
+				<Header />
 				<S.Content>
 					<CustomCalendar />
 					<AddTodo />
