+++ src/styles/todo/TodoMainPage.style.ts
@@ -6,6 +6,7 @@ export const Container = styled.div`
 	justify-content: space-between;
 	align-items: center;
 	height: 100vh;
+	padding: 30px 0;
 	overflow: hidden;
 `;
 
@@ -19,5 +20,5 @@ export const Content = styled.div`
 	width: 100%;
 	height: calc(100vh - 100px);
 	box-sizing: border-box;
-	margin-top: 50px;
+	margin-top: 70px;
 `;
